feat(finalize): consolidate merge + cleanup into vrg-finalize-pr with pre-merge provenance check
