+++ pyproject.toml
@@ -22,6 +22,7 @@ vrg-docs-stage = "vergil_tooling.bin.vrg_docs_stage:main"
 vrg-ecosystem-resolve = "vergil_tooling.bin.vrg_ecosystem_resolve:main"
 vrg-container-test = "vergil_tooling.bin.vrg_container_test:main"
 vrg-ensure-label = "vergil_tooling.bin.vrg_ensure_label:main"
+vrg-finalize-pr = "vergil_tooling.bin.vrg_finalize_pr:main"
 vrg-finalize-repo = "vergil_tooling.bin.vrg_finalize_repo:main"
 vrg-freeze-refs = "vergil_tooling.bin.vrg_freeze_refs:main"
 vrg-generate-commands = "vergil_tooling.bin.vrg_generate_commands:main"
